Breaking up llvm-ir2vec.cpp , extracting out tool classes to prepare a common importable module
